test.1b19/requirements.txt
Django>=3.2
djangorestframework
django-admirarchy
